+++ packages/matrix/src/heatmap/HeatMapCircle.tsx
@@ -1,7 +1,7 @@
 import { m } from 'framer-motion'
 import { HeatMapCellProps } from './types'
 
-export const HeatMapCircle = ({ x, y, width, height, className, style }: HeatMapCellProps) => {
+export const HeatMapCircle = ({ x, y, width, height, style, ...props }: HeatMapCellProps) => {
     const config = { fill: style?.fill, r: Math.min(width, height) / 2 }
     return (
         <m.circle
@@ -10,7 +10,7 @@ export const HeatMapCircle = ({ x, y, width, height, className, style }: HeatMap
             cx={x}
             cy={y}
             style={{ ...style, fill: undefined }}
-            className={className}
+            {...props}
         />
     )
 }