+++ packages/matrix/src/heatmap/HeatMapSimpleCircle.tsx
@@ -1,14 +1,7 @@
 import { HeatMapCellProps } from './types'
 
-export const HeatMapSimpleCircle = ({
-    x,
-    y,
-    width,
-    height,
-    className,
-    style,
-}: HeatMapCellProps) => {
-    return (
-        <circle cx={x} cy={y} r={Math.min(width, height) / 2} style={style} className={className} />
-    )
+export const HeatMapSimpleCircle = ({ x, y, width, height, ...props }: HeatMapCellProps) => {
+    delete props.cellSize
+    delete props.cellValue
+    return <circle cx={x} cy={y} r={Math.min(width, height) / 2} {...props} />
 }