+++ packages/matrix/src/heatmap/HeatMapSimpleRectangle.tsx
@@ -1,21 +1,7 @@
 import { HeatMapCellProps } from './types'
 
-export const HeatMapSimpleRectangle = ({
-    x,
-    y,
-    width,
-    height,
-    className,
-    style,
-}: HeatMapCellProps) => {
-    return (
-        <rect
-            x={x - width / 2}
-            y={y - height / 2}
-            height={height}
-            width={width}
-            style={style}
-            className={className}
-        />
-    )
+export const HeatMapSimpleRectangle = ({ x, y, width, height, ...props }: HeatMapCellProps) => {
+    delete props.cellSize
+    delete props.cellValue
+    return <rect x={x - width / 2} y={y - height / 2} height={height} width={width} {...props} />
 }